Addressing PR comments vol 1
