+++ DUI3-DX/Connectors/Rhino/Speckle.Connectors.Rhino7/Filters/RhinoEverythingFilter.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Speckle.Connectors.DUI.Bindings;
+using Speckle.Connectors.DUI.Models.Card.SendFilter;
 
 namespace Speckle.Connectors.Rhino7.Filters;
