+++ DUI3-DX/DUI3/Speckle.Connectors.DUI/Models/Card/CreateVersionArgs.cs
@@ -1,0 +1,7 @@
+﻿namespace Speckle.Connectors.DUI.Models.Card;
+
+public class CreateVersionArgs
+{
+  public string ModelCardId { get; set; }
+  public string ObjectId { get; set; }
+}
