+++ DUI3-DX/DUI3/Speckle.Connectors.DUI/Models/Card/ReceiveResult.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Speckle.Connectors.DUI.Models.Card;
+
+public class ReceiveResult
+{
+  public List<string> BakedObjectIds { get; set; }
+  public bool Display { get; set; }
+}
