+++ DUI3-DX/DUI3/Speckle.Connectors.DUI/Models/Card/ReceiverModelCard.cs
@@ -1,0 +1,9 @@
+namespace Speckle.Connectors.DUI.Models.Card;
+
+public class ReceiverModelCard : ModelCard
+{
+  public string ProjectName { get; set; }
+  public string ModelName { get; set; }
+  public string SelectedVersionId { get; set; }
+  public string LatestVersionId { get; set; }
+}
