+++ DUI3-DX/DUI3/Speckle.Connectors.DUI/Models/Card/ReceiverModelCardResult.cs
@@ -1,0 +1,7 @@
+namespace Speckle.Connectors.DUI.Models.Card;
+
+public class ReceiverModelCardResult
+{
+  public string ModelCardId { get; set; }
+  public ReceiveResult ReceiveResult { get; set; }
+}
