+++ DUI3-DX/DUI3/Speckle.Connectors.DUI/Models/Card/SenderModelCard.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+using Speckle.Connectors.DUI.Models.Card.SendFilter;
+using Speckle.Newtonsoft.Json;
+
+namespace Speckle.Connectors.DUI.Models.Card;
+
+public class SenderModelCard : ModelCard
+{
+  public ISendFilter SendFilter { get; set; }
+
+  [JsonIgnore]
+  public HashSet<string> ChangedObjectIds { get; set; } = new();
+}
